Python 3 patch for linode-python
Hi Ryan,

I needed to use the Linode API with Python 3 today, but
discovered some errors when installing it. I saw that you
have a Python 3 branch, but it has not been updated for a
few years, so I just made some minor fixes to the master
branch to make linode-python install cleanly (I am not sure
if it actually works with Python 3 otherwise). Please find
my patch attached.

--
Izaak

Signed-off-by: Ryan Tucker <rtucker@gmail.com>